+++ src/AuthenticationWithRolesModule.php
@@ -18,7 +18,13 @@
 
 namespace Rhubarb\Scaffolds\AuthenticationWithRoles;
 
+use Rhubarb\Custard\Command\CustardCommand;
 use Rhubarb\Scaffolds\Authentication\AuthenticationModule;
+use Rhubarb\Scaffolds\AuthenticationWithRoles\Commands\UpdateRolePermissionsCommand;
+use Rhubarb\Stem\Exceptions\RecordNotFoundException;
+use Rhubarb\Stem\Filters\Equals;
+use Rhubarb\Stem\Filters\Not;
+use Rhubarb\Stem\Filters\OneOf;
 use Rhubarb\Stem\Schema\SolutionSchema;
 
 /**
@@ -35,6 +41,78 @@ class AuthenticationWithRolesModule extends AuthenticationModule
     {
         parent::initialise();
 
-        SolutionSchema::registerSchema("Authentication", __NAMESPACE__ . '\DatabaseSchema');
+        SolutionSchema::registerSchema('Authentication', DatabaseSchema::class);
+    }
+
+    public static function updateRolePermissions()
+    {
+        /** @var Role[] $roleRecords */
+        $roleRecords = [];
+        /** @var Permission $permissionRecords */
+        $permissionRecords = [];
+
+        $permissionAssignmentIDs = [];
+
+        $permissionMaintainer = function (
+            $permissionRules,
+            $allow
+        ) use (
+            &$permissionRecords,
+            &$roleRecords,
+            &$permissionAssignmentIDs
+        ) {
+            foreach ($permissionRules as $roleName => $permissions) {
+                if (!isset($roleRecords[$roleName])) {
+                    try {
+                        $role = Role::findFirst(new Equals('RoleName', $roleName));
+                    } catch (RecordNotFoundException $ex) {
+                        $role = new Role();
+                        $role->RoleName = $roleName;
+                        $role->save();
+                    }
+                    $roleRecords[$roleName] = $role;
+                }
+
+                foreach ($permissions as $permissionPath) {
+                    if (!isset($permissionRecords[$permissionPath])) {
+                        try {
+                            $permission = Permission::findFirst(new Equals('PermissionPath', $permissionPath));
+                        } catch (RecordNotFoundException $ex) {
+                            $permission = new Permission();
+                            $permission->PermissionPath = $permissionPath;
+                            $permission->PermissionName = $permissionPath;
+                            $permission->save();
+                        }
+                        $permissionRecords[$permissionPath] = $permission;
+                    }
+                    $permissionAssignmentIDs[] = $allow
+                        ? $roleRecords[$roleName]->allow($permissionRecords[$permissionPath])
+                        : $roleRecords[$roleName]->deny($permissionRecords[$permissionPath]);
+                }
+            }
+        };
+
+        /** @var RolePermissionDefinitions $settings */
+        $settings = RolePermissionDefinitions::singleton();
+
+        $permissionMaintainer($settings->allowRolePermissions, true);
+        $permissionMaintainer($settings->denyRolePermissions, false);
+
+        if (count($permissionAssignmentIDs) > 0) {
+            PermissionAssignment::find(
+                new Not(new OneOf('PermissionAssignmentID', $permissionAssignmentIDs))
+            )->deleteAll();
+        }
+    }
+
+    /**
+     * @return CustardCommand[]
+     */
+    public function getCustardCommands()
+    {
+        $commands = parent::getCustardCommands();
+        $commands[] = new UpdateRolePermissionsCommand();
+
+        return $commands;
     }
 }