+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change log
 
+### 1.2.0
+
+* Added:    Codeception
+* Added:    .gitignore file
+* Added:    Custard command to update roles and permissions
+* Added:    Module now has method for updating roles and managing their assigned permissions
+
 ### 1.1.1
+
 * Fixed:    Authentication module updated with leaf v1 support
 
 ### 1.1.0